requirements.txt
Flask==3.0.1
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-datastore==2.20.0
